integrated the bracket display at the end of each round, and added game menu logo
