+++ frontend/src/pages/Game/game.css
@@ -266,3 +266,240 @@
         gap: 30px;
     }
 }
+
+/* ================================================================== */
+/* ====================== Tournament Bracket ======================== */
+/* ================================================================== */
+
+.game-touranment {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 100%;
+    height: 100vh;
+    max-height: 100%;
+    background-color: var(--base-shade);
+    z-index: 2001;
+    padding: 80px 5%;
+}
+
+.tournament-content {
+    display: flex;
+    flex-direction: column;
+    width: fit-content;
+    gap: 30px;
+}
+
+.rounds {
+    display: flex;
+	justify-content: space-between;
+    padding: 20px 40px;
+    text-align: center;
+}
+
+.rounds p {
+    font-size: clamp(0.8rem, 1.2vw, 2rem);
+    font-weight: bold;
+    margin: 0;
+}
+
+.round-one, .final-round, .round-two{
+    display: flex;
+    justify-content: center;
+    padding: 10px 30px;
+    width: clamp(120px, 100%, 300px);
+    background: var(--base);
+    height: fit-content;
+    text-align: center;
+    border-radius: 12px;
+    border: 2px solid var(--sub-color);
+}
+
+.final-round{
+    border: 2px solid var(--primary-color);
+}
+
+.bracket{
+    display: flex;
+    justify-content: center;
+    flex-direction: row;
+    padding: 20px 40px;
+    gap: 10px;
+}
+
+.separator svg {
+    height: 150px;
+    fill: var(--primary-color);
+}
+
+.match-one, .match-two {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    gap: 30px;
+}
+
+.match-final {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+}
+
+.tounament-match-card {
+    display: flex;
+    align-items: center;
+    background-color: var(--base);
+    height: 60px;
+    width: 100%;
+    border-radius: 8px;
+}
+
+.match-content {
+    display: flex;
+    align-items: center;
+    position: relative;
+    padding: 10px 15px;
+    border: 1px solid var(--sub-color);
+    height: 100%;
+    width: 100%;
+}
+
+.match-content svg {
+    position: absolute;
+    top: 0;
+    left: 0;
+}
+
+.match-one .match-content {
+    border-radius: 8px 0 0 8px;
+}
+
+.match-two .match-content {
+    border-radius: 0 8px 8px 0;
+}
+
+.match-final .match-content {
+    border-radius: 8px;
+}
+
+.match-content .avatar {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    object-fit: cover;
+}
+
+.match-username {
+    text-align: center;
+    width: 150px;
+    color: var(--title-color);
+    font-size: clamp(0.8rem, 1.7vw, 1.2rem);
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+}
+
+.bracket-match-score {
+    display: flex;
+    justify-content: center;
+    border: 1px solid var(--sub-color);
+    height: 100%;
+    padding: 15px;
+    width: 50px;
+}
+
+.bracket-match-score p {
+    text-align: center;
+    font-size: 18px;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    margin: 0;
+}
+
+.match-one .bracket-match-score {
+    border-radius: 0 8px 8px 0;
+    border-left: 0;
+}
+
+.match-two .bracket-match-score {
+    border-radius: 8px 0 0 8px;
+    border-right: 0;
+}
+
+.separator {
+    display: flex;
+    width: 100%;
+    height: fit-content;
+}
+
+@media (max-width: 1100px) {
+    .bracket {
+        flex-direction: column;
+        padding: 20px;
+        gap: 40px;
+    }
+
+    .separator{
+        display: none;
+    }
+
+    .separator svg {
+        height: 100px;
+        fill: var(--primary-color);
+        transform: rotate(90deg);
+    }
+
+    .match-two .bracket-match-score {
+        border-radius: 0px 8px 8px 0px;
+        border-left: 0;
+        border-right: 1px solid var(--sub-color);
+        order: 1;
+    }
+
+    .match-two .match-content {
+        border-radius: 8px 0px 0px 8px;
+    }
+
+    .match-final{
+        flex-direction: column;
+    }
+
+    .match-username {
+        width: 90px;
+        order: 1;
+    }
+
+    .tournament-content {
+        flex-direction: row;
+        gap: 0;
+    }
+
+    .round-one, .final-round, .round-two{
+        padding: 10px;
+        width: fit-content;
+    }
+
+    .rounds {
+        padding: 20px;
+        flex-direction: column;
+    }
+
+    .rounds p{
+        writing-mode: vertical-rl;
+        text-orientation: upright;
+    }
+}
+
+.title-img {
+    max-width: 70%;
+    max-height: 100%;
+    object-fit: cover;
+}
+
