+++ frontend/src/pages/Game/game.html
@@ -1,4 +1,6 @@
 <div class="main-menu">
+	<!-- add icon here -->
+	<img class="title-img" src="./images/pongtitle.png" alt="">
 	<button id="start" class="btn-game">1 VS 1</button>
 	<button id="tournament" class="btn-game">Tournament</button>
 	<div class="player-nick d-none" id="player-nick"> <!-- d-none -->
